fix: moved interface trace
