+++ src/services/types.ts
@@ -70,6 +70,20 @@ type DmDecisions =
   | "CREAR_OBLIGACIONES_DE_CARTERA"
   | "DECLINAR_OBLIGACIONES_DE_CARTERA";
 
+  interface TraceType {
+    trace_id: string;
+    trace_value: string;
+    credit_request_id: string;
+    use_case: string;
+    user_id: string;
+    execution_date: string | number;
+    justification?: string;
+    decision_taken_by_user?: string;
+    trace_type?: string;
+    read_novelty?: string;
+  }
+  
+
 export type {
   Requests,
   IStaff,
@@ -78,6 +92,7 @@ export type {
   DmConceptos,
   DmDecisions,
   PinnedRequest,
+  TraceType
 };
 export interface approval_by_credit_request_Mock {
   approval_id: string;
